Cleanup graphql tree types
