+++ packages/graphql/src/api-v6/queryIRFactory/resolve-tree-parser/ResolveTreeParser.ts
@@ -24,22 +24,22 @@ import type { Attribute } from "../../../schema-model/attribute/Attribute";
 import { ListType } from "../../../schema-model/attribute/AttributeType";
 import { ConcreteEntity } from "../../../schema-model/entity/ConcreteEntity";
 import type { Relationship } from "../../../schema-model/relationship/Relationship";
+
 import type {
-    GraphQLConnectionArgs,
-    GraphQLReadOperationArgs,
-    GraphQLSortArgument,
-    GraphQLSortEdgeArgument,
     GraphQLTreeCartesianPoint,
+    GraphQLTreeLeafField,
+    GraphQLTreePoint,
+    GraphQLTreeScalarField,
+} from "./graphql-tree/attributes";
+import type {
+    GraphQLTree,
     GraphQLTreeConnection,
     GraphQLTreeEdge,
     GraphQLTreeEdgeProperties,
-    GraphQLTreeLeafField,
     GraphQLTreeNode,
-    GraphQLTreePoint,
     GraphQLTreeReadOperation,
-    GraphQLTreeScalarField,
-    GraphQLTreeSortElement,
-} from "./graphql-tree";
+} from "./graphql-tree/graphql-tree";
+import type { GraphQLSortArgument, GraphQLSortEdgeArgument, GraphQLTreeSortElement } from "./graphql-tree/sort";
 import { findFieldByName } from "./utils/find-field-by-name";
 
 export abstract class ResolveTreeParser<T extends ConcreteEntity | Relationship> {
@@ -69,7 +69,7 @@ export abstract class ResolveTreeParser<T extends ConcreteEntity | Relationship>
         };
     }
 
-    private parseOperationArgs(resolveTreeArgs: Record<string, any>): GraphQLReadOperationArgs {
+    private parseOperationArgs(resolveTreeArgs: Record<string, any>): GraphQLTree["args"] {
         // Not properly parsed, assuming the type is the same
         return {
             where: resolveTreeArgs.where,
@@ -221,7 +221,7 @@ export abstract class ResolveTreeParser<T extends ConcreteEntity | Relationship>
         return relationshipTreeParser.parseOperation(resolveTree);
     }
 
-    private parseConnectionArgs(resolveTreeArgs: { [str: string]: any }): GraphQLConnectionArgs {
+    private parseConnectionArgs(resolveTreeArgs: { [str: string]: any }): GraphQLTreeConnection["args"] {
         let sortArg: GraphQLSortArgument[] | undefined;
         if (resolveTreeArgs.sort) {
             sortArg = resolveTreeArgs.sort.map((sortArg): GraphQLSortArgument => {
@@ -341,6 +341,5 @@ function resolveTreeToLeafField(resolveTree: ResolveTree | undefined): GraphQLTr
         alias: resolveTree.alias,
         args: resolveTree.args,
         name: resolveTree.name,
-        fields: undefined,
     };
 }
