+++ packages/graphql/src/api-v6/queryIRFactory/resolve-tree-parser/graphql-tree/graphql-tree.ts
@@ -1,0 +1,83 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { Integer } from "neo4j-driver";
+import type { GraphQLTreeLeafField } from "./attributes";
+import type { GraphQLSortArgument, GraphQLSortEdgeArgument } from "./sort";
+import type { GraphQLTreeElement } from "./tree-element";
+import type { GraphQLWhereArgs, GraphQLWhereArgsTopLevel } from "./where";
+
+export type GraphQLTree = GraphQLTreeReadOperationTopLevel;
+
+interface GraphQLTreeReadOperationTopLevel extends GraphQLTreeElement {
+    name: string;
+    fields: {
+        connection?: GraphQLTreeConnectionTopLevel;
+    };
+    args: {
+        where?: GraphQLWhereArgsTopLevel;
+    };
+}
+
+export interface GraphQLTreeReadOperation extends GraphQLTreeElement {
+    name: string;
+    fields: {
+        connection?: GraphQLTreeConnection;
+    };
+    args: {
+        where?: GraphQLWhereArgs;
+    };
+}
+
+export interface GraphQLTreeConnection extends GraphQLTreeElement {
+    fields: {
+        edges?: GraphQLTreeEdge;
+    };
+    args: {
+        sort?: GraphQLSortArgument[];
+        first?: Integer;
+        after?: string;
+    };
+}
+
+export interface GraphQLTreeConnectionTopLevel extends GraphQLTreeElement {
+    fields: {
+        edges?: GraphQLTreeEdge;
+    };
+    args: {
+        sort?: GraphQLSortEdgeArgument[];
+        first?: Integer;
+        after?: string;
+    };
+}
+
+export interface GraphQLTreeEdge extends GraphQLTreeElement {
+    fields: {
+        node?: GraphQLTreeNode;
+        properties?: GraphQLTreeEdgeProperties;
+    };
+}
+
+export interface GraphQLTreeNode extends GraphQLTreeElement {
+    fields: Record<string, GraphQLTreeLeafField | GraphQLTreeReadOperation>;
+}
+
+export interface GraphQLTreeEdgeProperties extends GraphQLTreeElement {
+    fields: Record<string, GraphQLTreeLeafField>;
+}
